staticmethod _measurement update in dataset.py
Adds the option to only read in specific files types f.e. ".mea" or ".csv". The classmethods read_mea , read_zip and read_csv are also changed so they automatically read in the specific extensions by default. This patch prevents the TypeError that occurs when passing a folder that has not only the specified file types.